requirements.txt
Django==4.0.4
djangorestframework==3.13.1
APScheduler==3.6.3
pandas==1.4.2
gspread==5.3.2
psycopg2-binary==2.9.3
requests==2.27.1
telegram-send==0.34

django-apscheduler==0.6.2

google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1


black==22.3.0

cachetools==4.2.2